requirements.txt
python-telegram-bot==20.7
psutil==5.9.8 
colorlog==6.9.0